+++ src/axis.spec.js
@@ -1,6 +1,6 @@
 import defaultLocale from 'd3-time-format/locale/en-US.json';
 
-import axis, { tickFormat, getBreakpointLabel } from './axis';
+import axis, { tickFormat } from './axis';
 
 const defaultConfig = {
     label: {
@@ -22,15 +22,10 @@ const defaultConfig = {
     numberDisplayedTicks: {
         extra: 12,
     },
-    breakpoints: {
-        small: 576,
-        medium: 768,
-        large: 992,
-        extra: 1200,
-    },
 };
 
 const defaultScale = d3.scaleTime();
+const defaultBreakpointLabel = 'extra';
 
 describe('Axis Tick Format', () => {
     it('should format date correctly depending of current granularity', () => {
@@ -59,21 +54,27 @@ describe('Axis', () => {
     it('should add a group with class "axis"', () => {
         const selection = d3.select('svg').data([[{}]]);
 
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
         expect(document.querySelector('.axis')).toBeTruthy();
     });
 
     it('should append only a single group regardless number of given data', () => {
         const selection = d3.select('svg').data([[{}, {}, {}]]);
 
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
         expect(document.querySelectorAll('.axis').length).toBe(1);
     });
 
     it('should be translated of `label.width` to the right', () => {
         const selection = d3.select('svg').data([[{}, {}, {}]]);
 
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
         expect(document.querySelector('.axis').getAttribute('transform')).toBe(
             'translate(200,0)'
         );
@@ -83,7 +84,9 @@ describe('Axis', () => {
         const selection = d3.select('svg').data([[{}, {}]]);
 
         const axisTopSpy = jest.spyOn(d3, 'axisTop');
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
         expect(axisTopSpy).toHaveBeenCalledWith(defaultScale);
     });
 
@@ -98,7 +101,9 @@ describe('Axis', () => {
 
         const data = [[{ id: 'foo' }]];
         const selection = d3.select('svg').data(data);
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
 
         const tickFormat = tickFormatSpy.mock.calls[0][0];
         expect(tickFormat(new Date('2017-01-01 12:00:00'))).toBe('12 PM');
@@ -107,10 +112,14 @@ describe('Axis', () => {
     it('should remove "axis" group when data is removed', () => {
         const data = [[{ id: 'foo' }]];
         const selection = d3.select('svg').data(data);
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
 
         selection.data([[]]);
-        axis(d3, defaultConfig, defaultScale)(selection);
+        axis(d3, defaultConfig, defaultScale, defaultBreakpointLabel)(
+            selection
+        );
         expect(document.querySelectorAll('.axis').length).toBe(0);
     });
 
@@ -127,7 +136,7 @@ describe('Axis', () => {
 
         const data = [[{ id: 'foo' }]];
         const selection = d3.select('svg').data(data);
-        axis(d3, config, defaultScale)(selection);
+        axis(d3, config, defaultScale, defaultBreakpointLabel)(selection);
 
         expect(timeFormatDefaultLocaleSpy).toHaveBeenCalledWith(defaultLocale);
     });
@@ -141,29 +150,12 @@ describe('Axis', () => {
             numberDisplayedTicks: { extra: 9 },
         };
 
-        axis(d3, config, defaultScale)(selection);
+        axis(d3, config, defaultScale, defaultBreakpointLabel)(selection);
         expect(document.querySelectorAll('.tick').length).toBe(9);
     });
 
     afterEach(() => {
         document.body.innerHTML = '';
         jest.restoreAllMocks();
-    });
-});
-
-describe('Breakpoint Label', () => {
-    it('should return breakpoint label correctly depending of current point', () => {
-        expect(getBreakpointLabel(defaultConfig.breakpoints, 400)).toBe(
-            'small'
-        );
-        expect(getBreakpointLabel(defaultConfig.breakpoints, 600)).toBe(
-            'medium'
-        );
-        expect(getBreakpointLabel(defaultConfig.breakpoints, 800)).toBe(
-            'large'
-        );
-        expect(getBreakpointLabel(defaultConfig.breakpoints, 1000)).toBe(
-            'extra'
-        );
     });
 });
