+++ src/index.js
@@ -1,6 +1,7 @@
 import defaultsDeep from 'lodash.defaultsdeep';
 
-import axis, { getBreakpointLabel } from './axis';
+import axis from './axis';
+import { getBreakpointLabel } from './breakpoint';
 import bounds from './bounds';
 import defaultConfiguration from './config';
 import dropLine from './dropLine';
@@ -46,7 +47,10 @@ export default ({ d3 = window.d3, ...customConfiguration }) => {
             .range([0, width - labelWidth]);
 
         chart._scale = xScale;
-        chart._breakpoint = getBreakpointLabel(breakpoints, window.innerWidth);
+        chart.currentBreakpointLabel = getBreakpointLabel(
+            breakpoints,
+            global.innerWidth
+        );
 
         const svg = root
             .enter()
@@ -126,7 +130,7 @@ export default ({ d3 = window.d3, ...customConfiguration }) => {
             .data(filteredData)
             .call(dropLine(config, scale))
             .call(bounds(config, scale))
-            .call(axis(d3, config, scale, chart._breakpoint));
+            .call(axis(d3, config, scale, chart.currentBreakpointLabel));
     };
 
     chart.draw = draw;
